+++ packages/react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @authon/react
 
+## 0.8.1
+
+- Adopt the updated shared refresh-session rotation contract.
+
 ## 0.8.0
 
 ### Minor Changes